pyproject.toml
[tool.poetry]
name = "dnd-maps"
version = "0.1.0"
description = ""
authors = ["Noah <noah@coronasoftware.net>"]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^6.0"
more-itertools = "^7.0"
numpy = "^1.16"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.7"

[tool.poetry.scripts]
start = "dnd_maps.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
